# server/public/swagger_output_v1.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Carved Rock Fitness API",
    "description": "API for Rest Fundamentals course on Pluralsight"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1"
    }
  ],
  "paths": {
    "/items/": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Gets all items",
        "description": "",
        "responses": {
          "200": {
            "description": "The list of items",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/items"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/items"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Creates a new item",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/itemDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/itemDTO"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Updates an item",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/updateItemDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/updateItemDTO"
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Gets a specific item by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/itemDetailed"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/itemDetailed"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Deletes a specific item by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The item that was deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/item"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/item"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/customers/": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Gets all customers",
        "description": "",
        "responses": {
          "200": {
            "description": "The list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/customers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/customers"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Creates a new customer",
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/createCustomerDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/createCustomerDTO"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "summary": "Updates a customer",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/updateCustomerDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/updateCustomerDTO"
              }
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Gets a specific customer by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/customer"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/customer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Customers"
        ],
        "summary": "Deletes a specific customer by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The customer that was deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/customer"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/customer"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/customers/{id}/orders": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Gets the orders for a customer by customer ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/customerOrders"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/customerOrders"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/customers/search/{query}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Gets customers matching the query",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/customers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/customers"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/orders/{id}/items/{itemId}": {
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "Deletes a specific item from an order by order and item ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order after the item is deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#components/schemas/order"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#components/schemas/order"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/orders/{id}/items": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Adds items to an order",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#components/schemas/orderItemsDTO"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#components/schemas/orderItemsDTO"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}